commit refactor viewmodels
