+++ bookShopSolution/bookShopSolution.Application/Catalog/Products/IPublicProductService.cs
@@ -1,5 +1,4 @@
 ﻿using bookShopSolution.ViewModels.Catalog.Products;
-using bookShopSolution.ViewModels.Catalog.Products.Public;
 using bookShopSolution.ViewModels.common;
 using System;
 using System.Collections.Generic;
@@ -11,6 +10,6 @@ namespace bookShopSolution.Application.Catalog.Products
 {
     public interface IPublicProductService
     {
-        Task<PagedResult<ProductViewModel>> GetAllByCategory(GetProductPagingRequest request);
+        Task<PagedResult<ProductViewModel>> GetAllByCategory(GetPublicProductPagingRequest request);
     }
 }
