+++ bookShopSolution/bookShopSolution.ViewModels/Catalog/Products/GetManageProductPagingRequest.cs
@@ -5,9 +5,9 @@ using System.Linq;
 using System.Text;
 using System.Threading.Tasks;
 
-namespace bookShopSolution.ViewModels.Catalog.Products.Manage
+namespace bookShopSolution.ViewModels.Catalog.Products
 {
-    public class GetProductPagingRequest : PagingRequestBase
+    public class GetManageProductPagingRequest : PagingRequestBase
     {
         public string Keyword { get; set; }
         public List<int> CategoryIds { get; set; }
