+++ bookShopSolution/bookShopSolution.ViewModels/Catalog/Products/GetPublicProductPagingRequest.cs
@@ -5,9 +5,9 @@ using System.Linq;
 using System.Text;
 using System.Threading.Tasks;
 
-namespace bookShopSolution.ViewModels.Catalog.Products.Public
+namespace bookShopSolution.ViewModels.Catalog.Products
 {
-    public class GetProductPagingRequest : PagingRequestBase
+    public class GetPublicProductPagingRequest : PagingRequestBase
     {
         public int? CategoryId { get; set; }
     }
