+++ bookShopSolution/bookShopSolution.ViewModels/Catalog/Products/ProductCreateRequest.cs
@@ -1,14 +1,17 @@
-﻿using System;
+﻿using Microsoft.AspNetCore.Http;
+using System;
 using System.Collections.Generic;
 using System.Linq;
 using System.Text;
 using System.Threading.Tasks;
 
-namespace bookShopSolution.ViewModels.Catalog.Products.Manage
+namespace bookShopSolution.ViewModels.Catalog.Products
 {
-    public class ProductUpdateRequest
+    public class ProductCreateRequest
     {
-        public int ProductId { get; set; }
+        public decimal Price { get; set; }
+        public decimal OriginalPrice { get; set; }
+        public int Stock { get; set; }
         public string ProductName { get; set; }
         public string Description { get; set; }
         public string Details { get; set; }
@@ -16,5 +19,6 @@ namespace bookShopSolution.ViewModels.Catalog.Products.Manage
         public string SeoTitle { get; set; }
         public int LanguageId { get; set; }
         public string SeoAlias { get; set; }
+        public IFormFile ThumbnailImage { get; set; }
     }
 }
