+++ bookShopSolution/bookShopSolution.ViewModels/Catalog/Products/ProductUpdateRequest.cs
@@ -5,13 +5,11 @@ using System.Linq;
 using System.Text;
 using System.Threading.Tasks;
 
-namespace bookShopSolution.ViewModels.Catalog.Products.Manage
+namespace bookShopSolution.ViewModels.Catalog.Products
 {
-    public class ProductCreateRequest
+    public class ProductUpdateRequest
     {
-        public decimal Price { get; set; }
-        public decimal OriginalPrice { get; set; }
-        public int Stock { get; set; }
+        public int ProductId { get; set; }
         public string ProductName { get; set; }
         public string Description { get; set; }
         public string Details { get; set; }
@@ -19,6 +17,6 @@ namespace bookShopSolution.ViewModels.Catalog.Products.Manage
         public string SeoTitle { get; set; }
         public int LanguageId { get; set; }
         public string SeoAlias { get; set; }
-        public IFormFile MyProperty { get; set; }
+        public IFormFile ThumbnailImage { get; set; }
     }
 }
